% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-coerce.R
\name{from_lidr}
\alias{from_lidr}
\title{Convert a lidR LAS object to parallax}
\usage{
from_lidr(las, platform = "aerial")
}
\arguments{
\item{las}{A \code{LAS} object from the \code{lidR} package.}

\item{platform}{Character string: "aerial", "terrestrial", "uav", or "unknown".}
}
\value{
A \code{px_cloud} (or platform subtype).
}
\description{
Convert a lidR LAS object to parallax
}
