% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-print.R
\name{print.px_cloud}
\alias{print.px_cloud}
\title{Print a parallax point cloud}
\usage{
\method{print}{px_cloud}(x, ...)
}
\arguments{
\item{x}{A \code{px_cloud} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a parallax point cloud
}
