% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{px_alpha_shape}
\alias{px_alpha_shape}
\title{Alpha shape surface reconstruction}
\usage{
px_alpha_shape(cloud, alpha = 1)
}
\arguments{
\item{cloud}{A \code{px_cloud}.}

\item{alpha}{Numeric — alpha parameter controlling concavity.}
}
\value{
A \code{mesh_result} object.
}
\description{
Alpha shape surface reconstruction
}
