% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{px_ball_pivot}
\alias{px_ball_pivot}
\title{Ball Pivoting Algorithm surface reconstruction}
\usage{
px_ball_pivot(cloud, radii = NULL)
}
\arguments{
\item{cloud}{A \code{px_cloud} with normals.}

\item{radii}{Numeric vector of ball radii, or \code{NULL} for auto-computation.}
}
\value{
A \code{mesh_result} object.
}
\description{
Ball Pivoting Algorithm surface reconstruction
}
