% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{px_crop}
\alias{px_crop}
\title{Crop to bounding box}
\usage{
px_crop(cloud, bbox)
}
\arguments{
\item{cloud}{A \code{px_cloud}.}

\item{bbox}{Numeric vector of length 6: (xmin, ymin, zmin, xmax, ymax, zmax).}
}
\value{
A \code{px_cloud} — cropped.
}
\description{
Crop to bounding box
}
