% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{px_metrics}
\alias{px_metrics}
\title{Compute point cloud metrics}
\usage{
px_metrics(cloud)
}
\arguments{
\item{cloud}{A \code{px_cloud}.}
}
\value{
A tibble with one row containing summary statistics.
}
\description{
Summary statistics including point count, density, elevation distribution,
and intensity statistics.
}
