% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{px_plot_registration}
\alias{px_plot_registration}
\title{Visualize a registration result}
\usage{
px_plot_registration(source, target, result)
}
\arguments{
\item{source}{A \code{px_cloud}.}

\item{target}{A \code{px_cloud}.}

\item{result}{An \code{icp_result}.}
}
\value{
Invisibly returns a ggplot object.
}
\description{
Visualize a registration result
}
