% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{px_plot_segmentation}
\alias{px_plot_segmentation}
\title{Visualize segmentation results}
\usage{
px_plot_segmentation(cloud, result)
}
\arguments{
\item{cloud}{A \code{px_cloud}.}

\item{result}{A \code{segmentation_result}.}
}
\value{
Invisibly returns a ggplot object.
}
\description{
Visualize segmentation results
}
