% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{px_random_downsample}
\alias{px_random_downsample}
\title{Random downsampling}
\usage{
px_random_downsample(cloud, fraction = 0.5, seed = NULL)
}
\arguments{
\item{cloud}{A \code{px_cloud}.}

\item{fraction}{Numeric between 0 and 1 — fraction to keep.}

\item{seed}{Integer or \code{NULL} — random seed.}
}
\value{
A \code{px_cloud} — downsampled.
}
\description{
Random downsampling
}
