% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{px_voxel_downsample}
\alias{px_voxel_downsample}
\title{Voxel grid downsampling}
\usage{
px_voxel_downsample(cloud, voxel_size = 0.05)
}
\arguments{
\item{cloud}{A \code{px_cloud}.}

\item{voxel_size}{Numeric — voxel edge length in coordinate units.}
}
\value{
A \code{px_cloud} — downsampled.
}
\description{
Voxel grid downsampling
}
