% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-coerce.R
\name{to_lidr}
\alias{to_lidr}
\title{Convert a parallax point cloud to lidR LAS}
\usage{
to_lidr(cloud)
}
\arguments{
\item{cloud}{A \code{px_cloud} object.}
}
\value{
A \code{LAS} object.
}
\description{
Convert a parallax point cloud to lidR LAS
}
